+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/controllers/HurricaneController.java
@@ -12,8 +12,10 @@ import edu.illinois.ncsa.incore.common.dao.ISpaceRepository;
 import edu.illinois.ncsa.incore.common.dao.IUserAllocationsRepository;
 import edu.illinois.ncsa.incore.common.dao.IUserFinalQuotaRepository;
 import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
+import edu.illinois.ncsa.incore.common.models.DemandDefinition;
 import edu.illinois.ncsa.incore.common.models.Space;
 import edu.illinois.ncsa.incore.common.utils.AllocationUtils;
+import edu.illinois.ncsa.incore.common.utils.DemandUtils;
 import edu.illinois.ncsa.incore.common.utils.UserGroupUtils;
 import edu.illinois.ncsa.incore.common.utils.UserInfoUtils;
 import edu.illinois.ncsa.incore.service.hazard.dao.IHurricaneRepository;
@@ -150,6 +152,15 @@ public class HurricaneController {
             .collect(Collectors.toList());
 
         return accessibleHurricanes;
+    }
+
+    @GET
+    @Path("/demands")
+    @Produces({MediaType.APPLICATION_JSON})
+    @Operation(summary = "Returns all hurricane allowed demand types and units.")
+    public List<DemandDefinition> getHurricaneDemands() {
+        JSONObject demandDefinition = new JSONObject(commonRepository.getAllDemandDefinitions().get(0).toJson());
+        return DemandUtils.getAllowedDemands(demandDefinition, "hurricane");
     }
 
     @GET